==> rmds/2-run_estimators.Rmd <==
---
title: "2-Run estimators"
author: "Siyuan Ma"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  html_document:
    df_print: paged
---
# Overview
- Run five different estimators on the simulated datasets.
1. Naive Gehan estimator (with only fully observed data).
2. Combined Gehan estimator (including studies with systematically missing covariates).
3. Optimal Gehan estimator (optimal weighting of the two estimators).
4. Bivariate MLE estimator given in 19222087.
5. Multiple imputation estimator given in 23857554.
```{r setup, echo=FALSE} 
knitr::opts_knit$set(root.dir = normalizePath(".."))
```
```{r setup2, echo=FALSE}
rm(list = ls())
smar::sourceDir("functions/")
# dir_project <- "~/Dropbox (Harvard University)/Survival Project/survival_paper/"
dir_project <- "/n/hutlab11_nobackup/users/syma/survival_project/survival_paper/"
dir_output <- paste0(dir_project, "results/2-run_estimators/")
dir.create(dir_output, recursive = TRUE, showWarnings = FALSE)
library(magrittr)
```
```{r load datasets}
load(paste0(dir_project, "results/1-simulate_datasets/scenarioI.RData"))
# load(paste0(dir_project, "results/1-simulate_datasets/scenarioII.RData"))
# load(paste0(dir_project, "results/1-simulate_datasets/scenarioIII.RData"))
# tb_sim <- rbind(tb_sim_I %>% dplyr::mutate(scenario = 1),
#                 tb_sim_II %>% dplyr::mutate(scenario = 2),
#                 tb_sim_III %>% dplyr::mutate(scenario = 3))
tb_sim <- tb_sim_I
```
```{r run estimators}
B <- 500
N <- nrow(tb_sim)
start.time <- Sys.time()
l_results_I <-  Rmpi::mpi.parLapply(
  seq_len(N),
  function(i,
           dir_output,
           tb_sim,
           B) {
    library(magrittr)
    library(micemd)
    # cat(i, "\n", file = paste0(dir_output, "progress.txt"), append = TRUE)
    i_sim <- tb_sim[i, ]
    pX <- nrow(i_sim$a[[1]])
    pZ <- ncol(i_sim$a[[1]])
    
    fit.obs <-
      AFTmeta::gehan.obs(y = i_sim$dataset[[1]]$logY,
                         delta = i_sim$dataset[[1]]$delta,
                         matX = i_sim$dataset[[1]][, 1:pX, drop = FALSE] %>% as.matrix,
                         matZ = i_sim$dataset[[1]][, (pX + 1):(pX + pZ), drop = FALSE] %>% as.matrix,
                         study = i_sim$dataset[[1]]$study,
                         missing = i_sim$dataset[[1]]$missing,
                         B = B,
                         ncores = 1)
    fit.full <-
      AFTmeta::gehan.full(y = i_sim$dataset[[1]]$logY,
                          delta = i_sim$dataset[[1]]$delta,
                          matX = i_sim$dataset[[1]][, 1:pX, drop = FALSE] %>% as.matrix,
                          matZ = i_sim$dataset[[1]][, (pX + 1):(pX + pZ), drop = FALSE] %>% as.matrix,
                          study = i_sim$dataset[[1]]$study,
                          missing = i_sim$dataset[[1]]$missing,
                          B = B,
                          ncores = 1)
    fit.opt <-
      AFTmeta::gehan.opt(y = i_sim$dataset[[1]]$logY,
                         delta = i_sim$dataset[[1]]$delta,
                         matX = i_sim$dataset[[1]][, 1:pX, drop = FALSE] %>% as.matrix,
                         matZ = i_sim$dataset[[1]][, (pX + 1):(pX + pZ), drop = FALSE] %>% as.matrix,
                         study = i_sim$dataset[[1]]$study,
                         missing = i_sim$dataset[[1]]$missing,
                         B = B,
                         ncores = 1)
    fit.fib <-
      AFTmeta::gehan.fib(y = i_sim$dataset[[1]]$logY,
                         delta = i_sim$dataset[[1]]$delta,
                         matX = i_sim$dataset[[1]][, 1:pX, drop = FALSE] %>% as.matrix,
                         matZ = i_sim$dataset[[1]][, (pX + 1):(pX + pZ), drop = FALSE] %>% as.matrix,
                         study = i_sim$dataset[[1]]$study,
                         missing = i_sim$dataset[[1]]$missing,
                         B = B,
                         ncores = 1)
    fit.mi1 <-
      AFTmeta::gehan.mi(y = i_sim$dataset[[1]]$logY,
                        delta = i_sim$dataset[[1]]$delta,
                        matX = i_sim$dataset[[1]][, 1:pX, drop = FALSE] %>% as.matrix,
                        matZ = i_sim$dataset[[1]][, (pX + 1):(pX + pZ), drop = FALSE] %>% as.matrix,
                        study = i_sim$dataset[[1]]$study,
                        missing = i_sim$dataset[[1]]$missing,
                        m = 10) 
    fit.mi2 <-
      AFTmeta::gehan.mi(y = i_sim$dataset[[1]]$logY,
                        delta = i_sim$dataset[[1]]$delta,
                        matX = i_sim$dataset[[1]][, 1:pX, drop = FALSE] %>% as.matrix,
                        matZ = i_sim$dataset[[1]][, (pX + 1):(pX + pZ), drop = FALSE] %>% as.matrix,
                        study = i_sim$dataset[[1]]$study,
                        missing = i_sim$dataset[[1]]$missing,
                        surv_est = "marg",
                        surv_use = "H",
                        m = 10) 
    fit.mi3 <-
      AFTmeta::gehan.mi(y = i_sim$dataset[[1]]$logY,
                        delta = i_sim$dataset[[1]]$delta,
                        matX = i_sim$dataset[[1]][, 1:pX, drop = FALSE] %>% as.matrix,
                        matZ = i_sim$dataset[[1]][, (pX + 1):(pX + pZ), drop = FALSE] %>% as.matrix,
                        study = i_sim$dataset[[1]]$study,
                        missing = i_sim$dataset[[1]]$missing,
                        surv_est = "marg",
                        surv_use = "S",
                        m = 10) 
    fit.mi4 <-
      AFTmeta::gehan.mi(y = i_sim$dataset[[1]]$logY,
                        delta = i_sim$dataset[[1]]$delta,
                        matX = i_sim$dataset[[1]][, 1:pX, drop = FALSE] %>% as.matrix,
                        matZ = i_sim$dataset[[1]][, (pX + 1):(pX + pZ), drop = FALSE] %>% as.matrix,
                        study = i_sim$dataset[[1]]$study,
                        missing = i_sim$dataset[[1]]$missing,
                        surv_est = "cond",
                        surv_use = "H",
                        m = 10) 
    fit.mi5 <-
      AFTmeta::gehan.mi(y = i_sim$dataset[[1]]$logY,
                        delta = i_sim$dataset[[1]]$delta,
                        matX = i_sim$dataset[[1]][, 1:pX, drop = FALSE] %>% as.matrix,
                        matZ = i_sim$dataset[[1]][, (pX + 1):(pX + pZ), drop = FALSE] %>% as.matrix,
                        study = i_sim$dataset[[1]]$study,
                        missing = i_sim$dataset[[1]]$missing,
                        surv_est = "cond",
                        surv_use = "S",
                        m = 10) 
    
    return(list(
      obs = fit.obs,
      full = fit.full,
      opt = fit.opt,
      fib = fit.fib,
      mi1 = fit.mi1,
      mi2 = fit.mi2,
      mi3 = fit.mi3,
      mi4 = fit.mi4,
      mi5 = fit.mi5
    ))
  },
  dir_output = dir_output,
  tb_sim = tb_sim,
  B = B)
print(Sys.time() - start.time)
save(l_results_I, file = paste0(dir_output, "fits_updated_I.RData"))
```
