==> rmds/3-evaluate_estimators.Rmd <==
---
title: "3-Evaluate estimators"
author: "Siyuan Ma"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  html_document:
    df_print: paged
---
# Overview
- We compare the performance of five Gehan-type estimators in the context of meta-analuysi, where some studies might have systematically missing covariates.
  1. obs: the naive estimator, using only fully observed studies
  2. full: the combined estimator, using both fully observed studies as well as those with systematically missing covariates.
  3. opt: "optimal" estimator, as a weighted average of the first two estimators.
      - As can be seen from the results, the relative performance between obs and full can vary given underlying parameters. However opt should always have close performance as the better one of the two.
  4. fib: AFT version of the proposed bivariate normal MLE Cox PH estimator from 19222087.
      - This estimator is based on the observation that betahat (the fully adjusted regression coefficients, can only be fitted on non-missing studies), and gammahat (the marginal regression coefficients, can be fitted on both non-missing studies and studies with systematically missing covariates), should follow a joint (asymptotic) bivariate normal distribution. If one is willing to assume that the covariance of this joint normal is the same across studies, it is then possible to write the likelihood for all per-study betahats and gammahats as the sum of the full joint likelihood for (betahat, gammahat) on non-missing studies, and the marginal likelihood for gammahat alone on studies with missingness. 
      - I think the assumptions of this method fail when different cohort might have difference characterstics, resulting in different covariance structures on (betahat, gammahat) across studies. In practice, I tried to realize this by specifying different distributions of the error term for time, e_T, as well as different censoring percentages. However as the summary tables show they do not seem to affect its performance too much, with one exception (see Scenario II).
  5. mi: AFT version of the proposed multiple imputation estimator from 23857554.
      - They specified a MI model with random effects. In my implementation I used mice::mice and modelling the missing covariates after all other covariates, the censored log survival times, as well as the censoring variable.
      - Because data is generated with a linear model (log survival time ~ missing covariates + non-missing covariates), this means the MI would fail when censored survival time is not good surrogate for true survival. Additionally, any difference across cohorts (survival time distribution, censoring percentage) would cause this method to fail. Both of these can be observed across the three simulation scenarios.
      - 1/18/2020 update: I followed Michelle's advice and added in survival function value
      in the imuptation model. Five total imputation models are evaluated: 1. with only y, delta, X, and additionally 2. with marginal cumulative hazard function, 3. with marginal survival function, 4. with conditional cumulative hazard function, and 5. with conditional survival function.
- Tables here correspond to the three summary tables in the current version of the manuscript.

```{r setup, echo=FALSE} 
knitr::opts_knit$set(root.dir = normalizePath(".."))
```
```{r setup2, echo=FALSE}
rm(list = ls())
smar::sourceDir("functions/")
dir_project <- "~/Dropbox (Harvard University)/Survival Project/survival_paper/"
dir_output <- paste0(dir_project, "results/3-evaluate_estimators/")
dir.create(dir_output, recursive = TRUE, showWarnings = FALSE)
library(magrittr)
```
```{r load datasets}
load(paste0(dir_project, "results/1-simulate_datasets/scenarioI.RData"))
load(paste0(dir_project, "results/1-simulate_datasets/scenarioII.RData"))
load(paste0(dir_project, "results/1-simulate_datasets/scenarioIII.RData"))
load(paste0(dir_project, "results/2-run_estimators/fits_aggregated.RData"))
tb_sim <- rbind(
  tb_sim_I %>% dplyr::mutate(scenario = 1),
  tb_sim_II %>% dplyr::mutate(scenario = 2),
  tb_sim_III %>% dplyr::mutate(scenario = 3)
)
```

# Calculate metrics for all simulation scenarios.
Metrics include mean, standard error, bias, mean squared error, and coverage probability.

```{r evaluate estimators}
tb_eval <- tb_sim %>% 
  dplyr::mutate(results = l_results) %>% 
  dplyr::group_by(i_setup, scenario) %>% 
  dplyr::summarise(
    obs = list(evaluate(truth = b[[1]],
                        l_results = results %>% purrr::map("obs"))),
    full = list(evaluate(truth = b[[1]],
                         l_results = results %>% purrr::map("full"))),
    opt = list(evaluate(truth = b[[1]],
                        l_results = results %>% purrr::map("opt"))),
    fib = list(evaluate(truth = b[[1]],
                        l_results = results %>% purrr::map("fib"))),
    mi1 = list(evaluate(truth = b[[1]],
                        l_results = results %>% purrr::map("mi1"))),
    mi2 = list(evaluate(truth = b[[1]],
                        l_results = results %>% purrr::map("mi2"))),
    mi3 = list(evaluate(truth = b[[1]],
                        l_results = results %>% purrr::map("mi3"))),
    mi4 = list(evaluate(truth = b[[1]],
                        l_results = results %>% purrr::map("mi4"))),
    mi5 = list(evaluate(truth = b[[1]],
                        l_results = results %>% purrr::map("mi5")))
  ) %>% 
  dplyr::ungroup() %>% 
  dplyr::group_by(1:dplyr::n()) %>% 
  dplyr::mutate(result_summarised = rbind(
    obs[[1]] %>% dplyr::mutate(estimator = "obs"),
    full[[1]] %>% dplyr::mutate(estimator = "full"),
    opt[[1]] %>% dplyr::mutate(estimator = "opt"),
    fib[[1]] %>% dplyr::mutate(estimator = "fib"),
    mi1[[1]] %>% dplyr::mutate(estimator = "mi1"),
    mi2[[1]] %>% dplyr::mutate(estimator = "mi2"),
    mi3[[1]] %>% dplyr::mutate(estimator = "mi3"),
    mi4[[1]] %>% dplyr::mutate(estimator = "mi4"),
    mi5[[1]] %>% dplyr::mutate(estimator = "mi5")
  ) %>% 
    dplyr::mutate(i_setup = i_setup, scenario = scenario) %>% 
    list()) %>% 
  magrittr::extract2("result_summarised") %>% 
  purrr::reduce(rbind) %>% 
  dplyr::mutate(estimator = factor(estimator, levels = c("obs", 
                                                         "full", 
                                                         "opt", 
                                                         "fib", 
                                                         paste0("mi", 1:5))))
tb_eval <- tb_eval %>% 
  dplyr::left_join(tb_sim %>% dplyr::filter(rep == 1), by = c("i_setup", "scenario"))
```

# Scenario I: homogeneous studies
Here I simulated two studies with equal sample size (n=100), one with all 3 variabes and one with one covariate (Z) missing. beta for the non-missing covariates were held constant (1, 1), while beta for the missing covariate, b_Z varies (0, 1, 2). I also varied censoring percentages to either 0.2 for both studies or 0.4 for both.

- Main takeaways:
  - Relative performance of obs vs. full varies according to b_Z, but opt alwasy stays close to the better one.
      - This mainly impacts coefficients for non-missing variables; coefficient for Z does not have much difference in performance. This is intuitive as incorporating studies where Z is missing should have little impact on estimation for Z.
  - fib has similar performance as opt.
  - mi performs badly when censoring percentage is high.
  
```{r homogeneous}
tb_print <- tb_eval %>% 
  dplyr::filter(scenario == 1) %>% 
  dplyr::mutate(b_Z = b %>% purrr::map_dbl(3), 
                censoring = percCens %>% purrr::map_dbl(unique),
                setup = paste0("b_Z=", b_Z, ", ", censoring, " censoring")) %>% 
  dplyr::arrange(censoring, b_Z, variable, estimator) %>% 
  dplyr::select(setup, variable, estimator, mean, bias, sd_empirical, 
                sd_estimate, MSE, coverage)
```

Performance of estimates of the non-missing coefficients can vary given b_Z:

```{r homogeneous print b_M}
tb_print %>% 
  dplyr::filter(variable == 1) %>% 
  dplyr::select(-variable) %>% 
  knitr::kable(format = "html") %>% 
  kableExtra::kable_styling() %>% 
  kableExtra::scroll_box(height = "500px")
```

Not so much for Z itself:

```{r homogeneous print b_Z}
tb_print %>% 
  dplyr::filter(variable == 3) %>% 
  dplyr::select(-variable) %>% 
  knitr::kable(format = "html") %>% 
  kableExtra::kable_styling() %>% 
  kableExtra::scroll_box(height = "500px")
```

# Scenario II: studies with heterngeneous cohort characteristics
In this scenario the distribution of underlying survival time is different across cohorts, either in the family (normal or evd), or in its standard deviation. The interpretation would be that cohorts are of different populations and might have different characteristics. Assumptions for the fib estimator (that the asymptotic variance of (beta, gamma) is the same across studies) should fail here. Assumptions for the MI procedure also fails. Four studies (two missing, two non-missing) were simulated for each setup.

- Main takeaways:
  - The estimators obs, full, and opt are still unbiased because the objective function is calculated within each study then summed across. All assumptions for these estimating equation based estimators still hold.
  - fib also should not be biased, as it is estimated as betahat_avail + A( gammahat_avail - gammahat_missing), where betahat_avail and gammahat_avail are weighted average of per-study full and marginal coefficients in the fully observed studies. gammahat_missing is weighted average of marginal coefficients in studies with systematically missing covariates. 
      - I hoped to see scenarios where opt has better performance than fib, given that the latter's assumptions were violated. 
      - This seems to be only the case where the spread (i.e., standard deviation) of survival time changes greatly between studies.
  - MI estimator has very severe bias when cohort characteristics vary.

```{r heterogeneous e_T}
tb_print <- tb_eval %>% 
  dplyr::filter(scenario == 2) %>%
  dplyr::mutate(dist_eT = dist_eT %>% purrr::map_chr(paste, collapse = ","), 
                sd_eT = sd_eT %>% purrr::map_chr(paste, collapse = ","),
                setup = paste0("distribution ", dist_eT, ", sd(e_T) ", sd_eT)) %>% 
  dplyr::arrange(dist_eT, sd_eT, variable, estimator) %>% 
  dplyr::select(setup, variable, estimator, mean, bias, sd_empirical, sd_estimate, MSE, coverage)
tb_print %>% 
  dplyr::filter(variable == 1) %>% 
  dplyr::select(-variable) %>% 
  knitr::kable(format = "html") %>% 
  kableExtra::kable_styling() %>% 
  kableExtra::scroll_box(height = "500px")
```

# Scenario III: studies with heterngeneous study designs.
Lastly we evaluate the scenario where the censoring distribution varies across cohorts. The interpretation would be that cohorts have different designs (follow-up periods). Same as the previous scenario, assumptions for both the fib and the mi estimators fail here. 

- Main takeaways:
  - Largely the same as previous scenario. obs, full, opt, and fib are still unbiased. fib has good performance even though its assumptions failed. mi performs badly.

```{r heterogeneous percCens}
tb_print <- tb_eval %>% 
  dplyr::filter(scenario == 3) %>%
  dplyr::mutate(censoring = percCens %>% purrr::map_chr(paste, collapse = ","), 
                setup = paste0("censoring ", censoring)) %>% 
  dplyr::arrange(censoring, variable, estimator) %>% 
  dplyr::select(setup, variable, estimator, mean, bias, sd_empirical, sd_estimate, 
                MSE, coverage)
tb_print %>% 
  dplyr::filter(variable == 1) %>% 
  dplyr::select(-variable) %>% 
  knitr::kable(format = "html") %>% 
  kableExtra::kable_styling() %>% 
  kableExtra::scroll_box(height = "500px")
```
